refactor: implement feedback from PR review
