+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/analytics/SelectBrpScreenId.kt
@@ -1,0 +1,9 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics
+
+import uk.gov.onelogin.criorchestrator.libraries.analytics.ScreenId
+
+enum class SelectBrpScreenId(
+    override val rawId: String,
+) : ScreenId {
+    SelectBrp(rawId = "8349d732-fb9e-49f0-9b7c-5ed84ffe4613"),
+}
