+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/brp/SelectBrpAction.kt
@@ -1,0 +1,9 @@
+package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp
+
+sealed class SelectBrpAction {
+    data object NavigateToBrpConfirmation : SelectBrpAction()
+
+    data object NavigateToDrivingLicence : SelectBrpAction()
+
+    data object NavigateToTypesOfPhotoID : SelectBrpAction()
+}
