+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/brp/SelectBrpScreen.kt
@@ -1,5 +1,6 @@
 package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.brp
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
@@ -11,9 +12,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toPersistentList
 import uk.gov.android.ui.componentsv2.bulletedlist.BulletedListTitle
@@ -27,7 +26,6 @@ import uk.gov.android.ui.componentsv2.inputs.radio.RadioSelectionTitle
 import uk.gov.android.ui.componentsv2.warning.GdsWarningText
 import uk.gov.android.ui.patterns.leftalignedscreen.LeftAlignedScreen
 import uk.gov.android.ui.theme.m3.GdsTheme
-import uk.gov.android.ui.theme.spacingDouble
 import uk.gov.android.ui.theme.util.UnstableDesignSystemAPI
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.R
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi.nav.SelectDocumentDestinations
@@ -66,14 +64,7 @@ fun SelectBrpScreen(
     }
 
     SelectBrpScreenContent(
-        title = stringResource(SelectBrpConstants.titleId),
-        readMoreButtonText = stringResource(SelectBrpConstants.readMoreButtonTextId),
         onReadMoreClicked = viewModel::onReadMoreClicked,
-        selectionItems =
-            SelectBrpConstants.selectionItems
-                .map { stringResource(it) }
-                .toPersistentList(),
-        continueButtonText = stringResource(SelectBrpConstants.continueButtonTextId),
         onContinueClicked = viewModel::onContinueClicked,
         modifier = modifier,
     )
@@ -83,11 +74,7 @@ fun SelectBrpScreen(
 @OptIn(UnstableDesignSystemAPI::class)
 @Composable
 internal fun SelectBrpScreenContent(
-    title: String,
-    readMoreButtonText: String,
     onReadMoreClicked: () -> Unit,
-    selectionItems: ImmutableList<String>,
-    continueButtonText: String,
     onContinueClicked: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
@@ -95,13 +82,13 @@ internal fun SelectBrpScreenContent(
 
     LeftAlignedScreen(
         modifier = modifier,
-        title = {
+        title = { horizontalPadding ->
             GdsHeading(
-                title,
-                modifier = Modifier.padding(horizontal = spacingDouble),
+                stringResource(SelectBrpConstants.titleId),
+                modifier = Modifier.padding(horizontal = horizontalPadding),
             )
         },
-        body = {
+        body = { horizontalPadding ->
             item {
                 GdsBulletedList(
                     persistentListOf(
@@ -114,30 +101,35 @@ internal fun SelectBrpScreenContent(
                             text = stringResource(R.string.selectdocument_brp_bullet_title),
                             titleType = Text,
                         ),
-                    modifier = Modifier.padding(horizontal = spacingDouble),
+                    modifier = Modifier.padding(horizontal = horizontalPadding),
                 )
             }
 
             item {
                 GdsWarningText(
                     stringResource(R.string.selectdocument_brp_expiry),
-                    modifier = Modifier.padding(horizontal = spacingDouble),
+                    modifier = Modifier.padding(horizontal = horizontalPadding),
                 )
             }
 
             item {
                 GdsButton(
-                    text = readMoreButtonText,
+                    text = stringResource(SelectBrpConstants.readMoreButtonTextId),
                     buttonType = ButtonType.Secondary,
                     onClick = onReadMoreClicked,
-                    modifier = Modifier.padding(horizontal = 4.dp),
+                    modifier = Modifier.padding(horizontal = horizontalPadding),
                     textAlign = TextAlign.Start,
+                    contentPosition = Arrangement.Start,
+                    contentModifier = Modifier,
                 )
             }
 
             item {
                 GdsSelection(
-                    items = selectionItems,
+                    items =
+                        SelectBrpConstants.selectionItems
+                            .map { stringResource(it) }
+                            .toPersistentList(),
                     selectedItem = selectedItem,
                     onItemSelected = { selectedItem = it },
                     title =
@@ -150,7 +142,7 @@ internal fun SelectBrpScreenContent(
         },
         primaryButton = {
             GdsButton(
-                continueButtonText,
+                stringResource(SelectBrpConstants.continueButtonTextId),
                 buttonType = ButtonType.Primary,
                 onClick = {
                     selectedItem?.let {
@@ -169,15 +161,7 @@ internal fun SelectBrpScreenContent(
 internal fun PreviewSelectBrpScreen() {
     GdsTheme {
         SelectBrpScreenContent(
-            title = stringResource(SelectBrpConstants.titleId),
-            readMoreButtonText = stringResource(SelectBrpConstants.readMoreButtonTextId),
             {},
-            selectionItems =
-                persistentListOf(
-                    stringResource(SelectBrpConstants.selectionItems[0]),
-                    stringResource(SelectBrpConstants.selectionItems[1]),
-                ),
-            continueButtonText = stringResource(SelectBrpConstants.continueButtonTextId),
             {},
         )
     }
