+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/brp/SelectBrpViewModelModule.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewmodel.viewModelFactory
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectBrpAnalytics
+import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.analytics.SelectDocumentAnalytics
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Named
 
@@ -17,7 +17,7 @@ object SelectBrpViewModelModule {
 
     @Provides
     @Named(FACTORY_NAME)
-    fun provideFactory(analytics: SelectBrpAnalytics): ViewModelProvider.Factory =
+    fun provideFactory(analytics: SelectDocumentAnalytics): ViewModelProvider.Factory =
         viewModelFactory {
             initializer {
                 SelectBrpViewModel(analytics)
