+++ features/select-doc/internal/src/test/snapshots/images/uk.gov.onelogin.criorchestrator.features.selectdoc.internal_ScreenshotTest_screenshot[7.uk.gov.onelogin.criorchestrator.features.selectdoc.internal.confirmbrp.ConfirmBrpScreenKt_PreviewBrpConfirmation_]_cy.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:941fb1cf6b4d9b198912a5d90ba025a5784c666cf22bda40da54a2635196c270
+size 8893
