+++ features/select-doc/internal/src/test/snapshots/images/uk.gov.onelogin.criorchestrator.features.selectdoc.internal_ScreenshotTest_screenshot[8.uk.gov.onelogin.criorchestrator.features.selectdoc.internal.confirmbrp.ConfirmBrpScreenKt_PreviewBrpConfirmation_]_dark_night.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2da2cf68edd9b70b437f81c2e72ab4a1db7dfc830e491f8beb36861af47bb21a
+size 3118
